Split diagnostics JSON helper
